% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{checking}
\alias{checking}
\title{Internal checking function}
\usage{
checking(
  DynForest_obj = NULL,
  timeData,
  fixedData,
  idVar,
  timeVar,
  timeVarModel,
  Y,
  ntree = 200,
  mtry = 1,
  nodesize = 1,
  minsplit = 2,
  cause = 1
)
}
\arguments{
\item{DynForest_obj}{A \code{DynForest} object resulting from \code{DynForest()} function}

\item{timeData}{A data.frame containing the id and time measurements variables and the time-dependent predictors.}

\item{fixedData}{A data.frame containing the id variable and the time-fixed predictors. Non-continuous variables should be characterized as factor.}

\item{idVar}{A character indicating the name of variable to identify the subjects}

\item{timeVar}{A character indicating the name of time variable}

\item{timeVarModel}{A list for each time-dependent predictors containing a list of formula for fixed and random part from the mixed model}

\item{Y}{A list of output which should contain: \code{type} defines the nature of the output, can be "\code{surv}", "\code{scalar}" or "\code{factor}"; \code{Y} is the output variable; \code{id} is the vector of the identifiers for each individuals, they should be the same as the identifiers of the inputs.}

\item{ntree}{Number of trees to grow. Default value set to 200.}

\item{mtry}{Number of candidate variables randomly drawn at each node of the trees. This parameter should be tuned by minimizing the OOB error.}

\item{minsplit}{(Only with survival outcome) Minimal number of events required to split the node. Cannot be smaller than 2.}

\item{cause}{(Only with competing events) Number indicates the event of interest.}
}
\description{
Internal checking function
}
\keyword{internal}
