% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_times.R
\name{combine_times}
\alias{combine_times}
\title{Extend predictions for new times}
\usage{
combine_times(pred, newtimes, type = "surv")
}
\arguments{
\item{pred}{Prediction matrix}

\item{newtimes}{Numeric vector of new times}

\item{type}{Character indicating survival function (\code{type}="surv") or risk function (\code{type}="risk")}
}
\description{
Extend predictions for new times
}
\keyword{internal}
