% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simu2.R
\docType{data}
\name{data_simu2}
\alias{data_simu2}
\title{data_simu1 dataset}
\format{
Longitudinal dataset with 1200 rows and 13 columns for 200 subjects
\describe{
\item{id}{Subject identifier}
\item{time}{Time measurement}
\item{cont_covar1}{Continuous time-fixed predictor 1}
\item{cont_covar2}{Continuous time-fixed predictor 2}
\item{bin_covar1}{Binary time-fixed predictor 1}
\item{bin_covar2}{Binary time-fixed predictor 2}
\item{marker1}{Continuous time-dependent predictor 1}
\item{marker2}{Continuous time-dependent predictor 2}
\item{marker3}{Continuous time-dependent predictor 3}
\item{marker4}{Continuous time-dependent predictor 4}
\item{marker5}{Continuous time-dependent predictor 5}
\item{marker6}{Continuous time-dependent predictor 6}
\item{Y_res}{Continuous outcome}
}
}
\description{
Simulated dataset 2 with continuous outcome
}
\examples{
data(data_simu2)
}
