% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynForest-package.R
\docType{package}
\name{DynForest-package}
\alias{DynForest}
\alias{DynForest-package}
\title{DynForest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anthony Devaux \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
