% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fact_partitions.R
\name{Fact.partitions}
\alias{Fact.partitions}
\title{Factor partitions finder}
\usage{
Fact.partitions(Factor, id)
}
\arguments{
\item{id}{}
}
\description{
This function is used to find all the unique partitions of k factors into 2 groups
}
\keyword{internal}
