% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamMM.R
\name{getParamMM}
\alias{getParamMM}
\title{Function to update the list of parameters for each marker using those estimated from previous node}
\usage{
getParamMM(current_node, markers, params)
}
\arguments{
\item{current_node}{Current node of the tree}

\item{markers}{Character vector indicating the name of the markers to be updated}

\item{params}{List to be updated for the current node}
}
\value{
List with the updated parameters from the requested markers
}
\description{
Function to update the list of parameters for each marker using those estimated from previous node
}
\keyword{internal}
