% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impurity.R
\name{impurity}
\alias{impurity}
\title{Compute the impurity of a given vector}
\usage{
impurity(Y)
}
\arguments{
\item{Y}{Outcome data}
}
\description{
Compute the impurity of a given vector
}
\keyword{internal}
