% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impurity_split.R
\name{impurity_split}
\alias{impurity_split}
\title{Impurity Split}
\usage{
impurity_split(Y, split, cause = 1, randsplit = FALSE, splits_evt = NULL)
}
\arguments{
\item{Y}{Outcome data}

\item{split}{Vector containing the subjects groups}

\item{cause}{(Only with competing events) Number indicates the event of interest.}
}
\description{
Impurity Split
}
\keyword{internal}
