% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOB_rfshape.R
\name{OOB.rfshape}
\alias{OOB.rfshape}
\title{OOB for random Forest}
\usage{
OOB.rfshape(rf, Curve = NULL, Scalar = NULL, Factor = NULL,
  Shape = NULL, Image = NULL, Y, timeScale = 0.1, d_out = 0.1,
  splitrule = NULL)
}
\arguments{
\item{d_out}{}
}
\description{
OOB for random Forest
}
\keyword{internal}
