% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc2.R
\docType{data}
\name{pbc2}
\alias{pbc2}
\title{pbc2 dataset}
\format{
Two datasets containing: (1) longitudinal data (1945 rows and 14 variables) (2) survival data (312 rows and 6 variables)
\describe{
\item{id}{X}
\item{time}{X}
\item{ascites}{X}
\item{hepatomegaly}{X}
\item{spiders}{X}
\item{edema}{X}
\item{serBilir}{X}
\item{serChol}{X}
\item{albumin}{X}
\item{alkaline}{X}
\item{SGOT}{X}
\item{platelets}{X}
\item{prothrombin}{X}
\item{histologic}{X}
\item{drug}{X}
\item{age}{X}
\item{sex}{X}
\item{years}{X}
\item{event}{X}
}
}
\source{
pbc2 joineRML
}
\description{
pbc2 data from Mayo clinic
}
\examples{
data(pbc2)
}
