% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc2_long.R
\docType{data}
\name{pbc2_long}
\alias{pbc2_long}
\title{pbc2 dataset with longitudinal information}
\format{
A data frame with 1945 rows and 14 variables:
\describe{
\item{id}{}
\item{year}{}
\item{ascites}{}
\item{hepatomegaly}{}
\item{spiders}{}
\item{edema}{}
\item{serBilir}{}
\item{serChol}{}
\item{albumin}{}
\item{alkaline}{}
\item{SGOT}{}
\item{platelets}{}
\item{prothrombin}{}
\item{histologic}{}
}
}
\source{
pbc2 joineRML
}
\usage{
pbc2_long
}
\description{
pbc2 dataset with longitudinal information
}
\keyword{datasets}
