% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc2_surv.R
\docType{data}
\name{pbc2_surv}
\alias{pbc2_surv}
\title{pbc2 dataset with survival information}
\format{
A data frame with 312 rows and 6 variables:
\describe{
\item{id}{}
\item{drug}{}
\item{age}{}
\item{sex}{}
\item{years}{}
\item{event}{}
}
}
\source{
pbc2 joineRML
}
\usage{
pbc2_surv
}
\description{
pbc2 dataset with survival information
}
\keyword{datasets}
