% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_courbes.R
\name{permutation_courbes}
\alias{permutation_courbes}
\title{Permute courbes among the subjects}
\usage{
permutation_courbes(Courbes, id)
}
\arguments{
\item{Courbes}{Numeric vector of the marker}

\item{id}{Numeric vector of id}
}
\description{
Permute courbes among the subjects
}
\keyword{internal}
