% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nodeCIF.R
\name{plot_nodeCIF}
\alias{plot_nodeCIF}
\title{Plot the estimated Cumulative Incidence Functions (CIF) for given tree nodes}
\usage{
plot_nodeCIF(DynForest_obj, tree = NULL, nodes = NULL)
}
\arguments{
\item{DynForest_obj}{A DynForest object from \code{DynForest()} function}

\item{tree}{Integer indicating the tree identifier}

\item{nodes}{Identifiers for the selected nodes. Default is set to \code{NULL} displaying all tree nodes.}
}
\value{
Display the estimated CIF for given tree nodes
}
\description{
Plot the estimated Cumulative Incidence Functions (CIF) for given tree nodes
}
\examples{
\donttest{
data(pbc2)

# Get Gaussian distribution for longitudinal predictors
pbc2$serBilir <- log(pbc2$serBilir)
pbc2$SGOT <- log(pbc2$SGOT)
pbc2$albumin <- log(pbc2$albumin)
pbc2$alkaline <- log(pbc2$alkaline)

# Sample 100 subjects
set.seed(1234)
id <- unique(pbc2$id)
id_sample <- sample(id, 100)
id_row <- which(pbc2$id\%in\%id_sample)

pbc2_train <- pbc2[id_row,]

timeData_train <- pbc2_train[,c("id","time",
                                "serBilir","SGOT",
                                "albumin","alkaline")]

# Create object with longitudinal association for each predictor
timeVarModel <- list(serBilir = list(fixed = serBilir ~ time,
                                     random = ~ time),
                     SGOT = list(fixed = SGOT ~ time + I(time^2),
                                 random = ~ time + I(time^2)),
                     albumin = list(fixed = albumin ~ time,
                                    random = ~ time),
                     alkaline = list(fixed = alkaline ~ time,
                                     random = ~ time))

# Build fixed data
fixedData_train <- unique(pbc2_train[,c("id","age","drug","sex")])

# Build outcome data
Y <- list(type = "surv",
          Y = unique(pbc2_train[,c("id","years","event")]))

# Run DynForest function
res_dyn <- DynForest(timeData = timeData_train, fixedData = fixedData_train,
                     timeVar = "time", idVar = "id",
                     timeVarModel = timeVarModel, Y = Y,
                     ntree = 50, nodesize = 5, minsplit = 5,
                     cause = 2, ncores = 2, seed = 1234)

# Display CIF for nodes 224 and 225 from tree 1
plot_nodeCIF(DynForest_obj = res_dyn, tree = 1, nodes = c(224,225))
}
}
