% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.DynForest}
\alias{predict.DynForest}
\title{Prediction using dynamic random forests}
\usage{
\method{predict}{DynForest}(
  object,
  timeData = NULL,
  fixedData = NULL,
  idVar,
  timeVar,
  t0 = NULL,
  ...
)
}
\arguments{
\item{object}{\code{DynForest} or \code{DynForest_OOB} object containing the dynamic random forest used on train data}

\item{timeData}{A data.frame containing the id and time measurements variables and the time-dependent predictors.}

\item{fixedData}{A data.frame containing the id variable and the time-fixed predictors. Non-continuous variables should be characterized as factor.}

\item{idVar}{A character indicating the name of variable to identify the subjects}

\item{timeVar}{A character indicating the name of time variable}

\item{t0}{Landmark time}

\item{...}{Optional parameters to be passed to the low level function}
}
\value{
Return the outcome of interest for the new subjects: matrix of probability of event of interest in survival mode, average value in regression mode and most likely value in classification mode
}
\description{
Prediction using dynamic random forests
}
\examples{
\dontrun{
# Predict on subjects still at risk at landmark time at 4 years
id_pred <- unique(pbc2_pred$id[which(pbc2_pred$years>4)])
pbc2_pred <- pbc2_pred[which(pbc2_pred$id\%in\%id_pred),]
timeData_pred <- pbc2_pred[,c("id", "time", "serBilir", "SGOT", "albumin", "alkaline")]
fixedData_pred <- unique(pbc2_pred[,c("id","age","drug","sex")])

pred_dyn <- predict(object = res_dyn,
                    timeData = timeData_pred, fixedData = fixedData_pred,
                    idVar = "id", timeVar = "time",
                    t0 = 4)
}
}
