% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE.R
\name{predRE}
\alias{predRE}
\title{Function to compute individual random effects using hlme output parameters}
\usage{
predRE(model, formula, data)
}
\arguments{
\item{model}{output object from hlme function}

\item{formula}{list of formula for fixed and random part}

\item{data}{data to compute random effect}
}
\value{
A table of random-effects in column by subjects in row
}
\description{
Function to compute individual random effects using hlme output parameters
}
\keyword{internal}
