% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtmax.R
\name{Rtmax}
\alias{Rtmax}
\title{Randomized Dynamic tree}
\usage{
Rtmax(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y = NULL,
  mtry = 1,
  timeScale = 0.1,
  nsplit_option = "quantile",
  nodesize = 1
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Y}{[list]:}

\item{mtry}{[integer]:}

\item{timeScale}{[numeric]:}

\item{nodesize}{}
}
\description{
Randomized Dynamic tree
}
\keyword{internal}
