% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_depth.R
\name{var_depth}
\alias{var_depth}
\title{Extract characteristics from the trees building process}
\usage{
var_depth(DynForest_obj)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}
}
\value{
var_depth function return a list with the following elements:\tabular{ll}{
\code{min_depth} \tab A table providing for each feature in row: the average depth and the rank \cr
\tab \cr
\code{var_node_depth} \tab A table providing for each tree in column the minimal depth for each feature in row. NA indicates that the feature was not used for the corresponding tree \cr
\tab \cr
\code{var_count} \tab A table providing for each tree in column the number of times where the feature is used (in row). 0 value indicates that the feature was not used for the corresponding tree \cr
}
}
\description{
Extract characteristics from the trees building process
}
\examples{
\dontrun{

# Run var_depth function
res_varDepth <- var_depth(res_dyn)

}

}
\seealso{
\code{\link{DynForest}}
}
