% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_split.R
\name{var_split}
\alias{var_split}
\title{Split function to build the two daughter nodes}
\usage{
var_split(X, Y, nsplit_option = "quantile", nodesize = 1, init = NULL)
}
\arguments{
\item{X}{Input data}

\item{Y}{Outcome data}

\item{nsplit_option}{A character indicates how the values are chosen to build the two groups for the splitting rule (only for continuous predictors). Values are chosen using deciles (\code{nsplit_option}="quantile") or randomly (\code{nsplit_option}="sample"). Default value is "quantile".}

\item{nodesize}{Minimal number of subjects required in both child nodes to split. Cannot be smaller than 1.}

\item{init}{(Optional) Initial values for linear mixed models}
}
\description{
Split function to build the two daughter nodes
}
\keyword{internal}
