% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_split_factor.R
\name{var_split_factor}
\alias{var_split_factor}
\title{Split function to build the two daughter nodes from factor predictor}
\usage{
var_split_factor(
  X,
  Y,
  cause = 1,
  nodesize = 1,
  randsplit = FALSE,
  splits_evt = NULL
)
}
\arguments{
\item{X}{Input data}

\item{Y}{Outcome data}

\item{cause}{(Only with competing events) Number indicates the event of interest.}

\item{nodesize}{Minimal number of subjects required in both child nodes to split. Cannot be smaller than 1.}
}
\description{
Split function to build the two daughter nodes from factor predictor
}
\keyword{internal}
