% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_split_num.R
\name{var_split_num}
\alias{var_split_num}
\title{Split function to build the two daughter nodes from numeric predictors}
\usage{
var_split_num(
  X,
  Y,
  nsplit_option = "quantile",
  cause = 1,
  nodesize = 1,
  randsplit = FALSE,
  splits_evt = NULL
)
}
\arguments{
\item{X}{Input data}

\item{Y}{Outcome data}

\item{nsplit_option}{A character indicates how the values are chosen to build the two groups for the splitting rule (only for continuous predictors). Values are chosen using deciles (\code{nsplit_option}="quantile") or randomly (\code{nsplit_option}="sample"). Default value is "quantile".}

\item{cause}{(Only with competing events) Number indicates the event of interest.}

\item{nodesize}{Minimal number of subjects required in both child nodes to split. Cannot be smaller than 1.}
}
\description{
Split function to build the two daughter nodes from numeric predictors
}
\keyword{internal}
