% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gVIMP.R
\name{compute_gVIMP}
\alias{compute_gVIMP}
\title{Compute the grouped importance of variables (gVIMP) statistic}
\usage{
compute_gVIMP(DynForest_obj, group = NULL, ncores = NULL)
}
\arguments{
\item{DynForest_obj}{\code{DynForest_OOB} object containing the dynamic random forest used on train data}

\item{group}{A list of groups with the name of the predictors assigned in each group}

\item{ncores}{Number of cores used to grow trees in parallel. Default value is the number of cores of the computer-1.}
}
\value{
compute_OOBerror() function return a list with the following elements:\tabular{ll}{
\code{data} \tab A list containing the data used to grow the trees \cr
\tab \cr
\code{rf} \tab A table with each tree in column. Provide multiple charactistics about the tree building \cr
\tab \cr
\code{type} \tab Outcome type \cr
\tab \cr
\code{times} \tab A numeric vector containing the time-to-event for all subjects \cr
\tab \cr
\code{cause} \tab Indicating the cause of interest \cr
\tab \cr
\code{causes} \tab A numeric vector containing the causes indicator \cr
\tab \cr
\code{Inputs} \tab A list of 3 elements: \code{Curve}, \code{Scalar} and \code{Factor}. Each element contains the names of the predictors \cr
\tab \cr
\code{Curve.model} \tab A list of longitudinal markers containing the formula used for modeling in the random forest \cr
\tab \cr
\code{param} \tab A list containing the hyperparameters \cr
\tab \cr
\code{xerror} \tab A numeric vector containing the OOB error for each tree \cr
\tab \cr
\code{oob.err} \tab A numeric vector containing the OOB error for each subject \cr
\tab \cr
\code{oob.pred} \tab Outcome prediction for all subjects \cr
\tab \cr
\code{IBS.range} \tab A vector containing the IBS min and max \cr
\tab \cr
\code{gVIMP} \tab A numeric vector containing the gVIMP for each group defined in \code{group} argument \cr
}
}
\description{
Compute the grouped importance of variables (gVIMP) statistic
}
\examples{
\dontrun{
# Compute gVIMP statistic
res_dyn_VIMP <- compute_gVIMP(DynForest_obj = res_dyn_OOB,
                              group = list(group1 = c("serBilir","SGOT"),
                                           group2 = c("albumin","alkaline")))
}
}
\author{
Anthony Devaux (\email{anthony.devaux@u-bordeaux.fr})
}
