% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_VIMP.R
\name{compute_VIMP}
\alias{compute_VIMP}
\title{Compute the importance of variables (VIMP) statistic}
\usage{
compute_VIMP(DynForest_obj, ncores = NULL)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object containing the dynamic random forest used on train data}

\item{ncores}{Number of cores used to grow trees in parallel. Default value is the number of cores of the computer-1.}
}
\value{
\code{compute_OOBerror()} function return a list with the following elements:\tabular{ll}{
\code{data} \tab A list containing the data used to grow the trees \cr
\tab \cr
\code{rf} \tab A table with each tree in column. Provide multiple charactistics about the tree building \cr
\tab \cr
\code{type} \tab Outcome type \cr
\tab \cr
\code{times} \tab A numeric vector containing the time-to-event for all subjects \cr
\tab \cr
\code{cause} \tab Indicating the cause of interest \cr
\tab \cr
\code{causes} \tab A numeric vector containing the causes indicator \cr
\tab \cr
\code{Inputs} \tab A list of 3 elements: \code{Curve}, \code{Scalar} and \code{Factor}. Each element contains the names of the predictors \cr
\tab \cr
\code{Curve.model} \tab A list of longitudinal markers containing the formula used for modeling in the random forest \cr
\tab \cr
\code{param} \tab A list containing the hyperparameters \cr
\tab \cr
\code{xerror} \tab A numeric vector containing the OOB error for each tree \cr
\tab \cr
\code{oob.err} \tab A numeric vector containing the OOB error for each subject \cr
\tab \cr
\code{oob.pred} \tab Outcome prediction for all subjects \cr
\tab \cr
\code{IBS.range} \tab A vector containing the IBS min and max \cr
\tab \cr
\code{Importance} \tab A list of 3 elements: \code{Curve}, \code{Scalar} and \code{Factor}. Each element contains a numeric vector of VIMP statistic predictor in \code{Inputs} value \cr
}
}
\description{
Compute the importance of variables (VIMP) statistic
}
\examples{
\dontrun{
# Compute VIMP statistic
res_dyn_VIMP <- compute_VIMP(DynForest_obj = res_dyn_OOB)
}
}
\author{
Anthony Devaux (\email{anthony.devaux@u-bordeaux.fr})
}
