% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vimp2.R
\name{compute_vimp2}
\alias{compute_vimp2}
\title{Compute the importance of variables (VIMP) statistic}
\usage{
compute_vimp2(
  dynforest_obj,
  IBS.min = 0,
  IBS.max = NULL,
  ncores = NULL,
  seed = 1234,
  permute_trajectory = FALSE
)
}
\arguments{
\item{dynforest_obj}{dynforest object resulting from \code{dynforest()}}

\item{IBS.min}{(Only with survival outcome) Minimal time to compute the Integrated Brier Score. Default is 0.}

\item{IBS.max}{(Only with survival outcome) Maximal time to compute the Integrated Brier Score. Default is the maximal time-to-event found.}

\item{ncores}{Number of cores used to grow trees in parallel. Default is the number of available cores minus 1.}

\item{seed}{Seed for reproducibility}

\item{permute_trajectory}{Logical. If TRUE, permutes longitudinal variables by exchanging patient trajectories instead of shuffling raw values. Default is FALSE.}
}
\value{
A list with VIMP results, including \code{Inputs}, \code{Importance}, \code{tree_oob_err}, and \code{IBS.range}.
}
\description{
Compute the importance of variables (VIMP) statistic
}
\seealso{
\code{\link[=dynforest]{dynforest()}}
}
