% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynForest.R
\name{DynForest}
\alias{DynForest}
\title{Joint random forest for longitudinal and survival data}
\usage{
DynForest(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  mtry = NULL,
  ntree = 100,
  ncores = NULL,
  timeScale = 0.1,
  imp = TRUE,
  d_out = 0.1,
  nsplit_option = "quantile",
  nodesize = 1,
  minsplit = 2,
  cause = 1,
  IBS.min = 0,
  IBS.max = NULL,
  seed = round(runif(1, 0, 10000)),
  ...
)
}
\arguments{
\item{Curve}{[list]: A list of longitudinal predictors which should contain: \code{X} a dataframe with one row for repeated measurement and as many columns as markers; \code{id} is the vector of the identifiers for the repeated measurements contained in \code{X}; \code{time} is the vector of the measurement times contained in \code{X}.}

\item{Scalar}{[list]: A list of scalar predictors which should contain: \code{X} a dataframe with as many columns as scalar predictors; \code{id} is the vector of the identifiers for each individual.}

\item{Factor}{[list]: A list of factor predictors which should contain: \code{X} a dataframe with as many columns as factor predictors; \code{id} is the vector of the identifiers for each individual.}

\item{Y}{[list]: A list of output which should contain: \code{type} defines the nature of the output, can be "\code{surv}", "\code{curve}", "\code{scalar}" or "\code{factor}"; \code{Y} is the output variable; \code{id} is the vector of the identifiers for each individuals, they should be the same as the identifiers of the inputs.}

\item{mtry}{[numeric]: Number of variables randomly drown as candidates at each split. The default value \code{p/3} but has to be tuned according to the OOB prediction error.}

\item{ntree}{[numeric]: Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}

\item{ncores}{[numeric]: Number of cores used to build Frechet randomized trees in parallel, defaulting to number of cores of the computer minus 1.}

\item{timeScale}{[numeric]: Allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping. Only used when there are trajectories either in input or output.}

\item{imp}{[logical]: TRUE to compute the variables importance FALSE otherwise (default \code{imp=}TRUE)}

\item{d_out}{[string]: "euc" or "frec".}

\item{nodesize}{[numeric]}

\item{minsplit}{[numeric]}

\item{...}{: optional parameters to be passed to the low level function}
}
\value{
A Frechet random forest which is a list of the following elements: \itemize{
\item \code{rf:} a list of the \code{ntree} randomized Frechet trees that compose the forest.
\item \code{xerror :} a vector containing the OOB prediction error of each randomized Frechet tree composing the forest.
\item \code{OOB.err: } a vector containing the OOB prediction error of each individual in the learning sample.
\item \code{OOB.pred: } a list of the OOB prediction for each individual in the learning set.
\item \code{Importance: } A vector containing the variables importance.
\item \code{varex: } “pseudo R-squared”: Percentage of variance explained.
}
}
\description{
DynForest function builds a joint random forest for survival analysis (Devaux et al., 2021), trajectory, regression or classification. Longitudinal data, factors and scalars are allowed as predictors.
Nodes are split on the candidate variable that maximize the splitting rule according to the outcome. To allow longitudinal data as candidate variable, mixed models are computed on those which are selected at each node.
Then, the random-effects are used are candidate variables. Out-of-bag prediction error and variable importance (VIMP) are also provided.
}
