% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynForest.R
\name{DynForest}
\alias{DynForest}
\title{Dynamic Random Forest}
\usage{
DynForest(Curve = NULL, Scalar = NULL, Factor = NULL, Y,
  mtry = NULL, ntree = 100, ncores = NULL, timeScale = 0.1,
  imp = TRUE, d_out = 0.1, nsplit_option = "quantile",
  nodesize = 1, cause = 1, IBS.min = 0, IBS.max = NULL, ...)
}
\arguments{
\item{Curve}{[list]: A list that contains the different input curves. It must contain the following elements (no choice): \code{X} the matrix of the different curves, each column code for a different curve variable; \code{id} is the vector of the identifiers for the different trajectories contained in \code{X}; \code{time} is the vector of the measurement times associated with the trajectories contained in \code{X}.}

\item{Scalar}{[list]: A list that contains the different input scalars. It must contain the following elements (no choice):  \code{X} the matrix of the scalars, each column code for a different variable; \code{id} is the vector of the identifiers for each individual.}

\item{Factor}{[list]: A list that contains the different input factors. It must contain the following elements (no choice):  \code{X} the matrix of the factors, each column code for a different variable; \code{id} is the vector of the identifiers for each individual.}

\item{Y}{[list]: A list that contains the output, It must contain the following elements (no choice): \code{type} defines the nature of the output, can be "\code{curve}", "\code{sclalar}", "\code{factor}", "\code{shape}", "\code{image}"; \code{Y} is the output variable; \code{id} is the vector of the identifiers for each individuals, they should be the same as the identifiers of the inputs.}

\item{mtry}{[numeric]: Number of variables randomly sampled as candidates at each split. The default value \code{p/3}}

\item{ntree}{[numeric]: Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}

\item{ncores}{[numeric]: Number of cores used to build Frechet randomized trees in parallel, defaulting to number of cores of the computer minus 1.}

\item{timeScale}{[numeric]: Allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping. Only used when there are trajectories either in input or output.}

\item{imp}{[logical]: TRUE to compute the variables importance FALSE otherwise (default \code{imp=}TRUE)}

\item{d_out}{[string]: "euc" or "frec".}

\item{...}{: optional parameters to be passed to the low level function}
}
\value{
A Frechet random forest which is a list of the following elements: \itemize{
\item \code{rf:} a list of the \code{ntree} randomized Frechet trees that compose the forest.
\item \code{xerror :} a vector containing the OOB prediction error of each randomized Frechet tree composing the forest.
\item \code{OOB.err: } a vector containing the OOB prediction error of each individual in the learning sample.
\item \code{OOB.pred: } a list of the OOB prediction for each individual in the learning set.
\item \code{Importance: } A vector containing the variables importance.
\item \code{varex: } “pseudo R-squared”: Percentage of variance explained.
}
}
\description{
This function builds Frechet random Forest introduced by Capitaine et.al, this includes the OOB predictions, OOB errors and variable importance computations.
}
