% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynForest.R
\name{DynForest}
\alias{DynForest}
\title{Random survival forest with multivariate longitudinal endogenous covariates}
\usage{
DynForest(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  mtry = NULL,
  ntree = 200,
  ncores = NULL,
  timeScale = 0.1,
  OOB_error = TRUE,
  imp = FALSE,
  imp.group = NULL,
  d_out = 0.1,
  nsplit_option = "quantile",
  nodesize = 1,
  minsplit = 2,
  cause = 1,
  IBS.min = 0,
  IBS.max = NULL,
  seed = round(runif(1, 0, 10000)),
  ...
)
}
\arguments{
\item{Curve}{A list of longitudinal predictors which should contain: \code{X} a dataframe with one row for repeated measurement and as many columns as markers; \code{id} is the vector of the identifiers for the repeated measurements contained in \code{X}; \code{time} is the vector of the measurement times contained in \code{X}.}

\item{Scalar}{A list of scalar predictors which should contain: \code{X} a dataframe with as many columns as scalar predictors; \code{id} is the vector of the identifiers for each individual.}

\item{Factor}{A list of factor predictors which should contain: \code{X} a dataframe with as many columns as factor predictors; \code{id} is the vector of the identifiers for each individual.}

\item{Y}{A list of output which should contain: \code{type} defines the nature of the output, can be "\code{surv}", "\code{curve}", "\code{scalar}" or "\code{factor}"; \code{Y} is the output variable; \code{id} is the vector of the identifiers for each individuals, they should be the same as the identifiers of the inputs.}

\item{mtry}{Number of candidate variables randomly drawn at each node of the trees. This parameter should be tuned by minimizing the OOB error. Default is `NULL`.}

\item{ntree}{Number of trees to grow. Default value set to 200.}

\item{ncores}{Number of cores used to grow trees in parallel. Default value is the number of cores of the computer-1.}

\item{timeScale}{Experimental}

\item{OOB_error}{Compute the OOB error. Default value is TRUE.}

\item{imp}{Compute (1) the importance of variables (VIMP); (2) the importance of groups (gVIMP) if there is a \code{imp.group} argument. Default value is FALSE}

\item{imp.group}{A list of groups with the name of the predictors assigned in each group}

\item{d_out}{Experimental}

\item{nsplit_option}{A character indicates how the values are chosen to build the two groups for the splitting rule (only for continuous predictors). Values are chosen using deciles (\code{nsplit_option}="quantile") or randomly (\code{nsplit_option}="sample"). Default value is "quantile".}

\item{nodesize}{Minimal number of subjects required in both child nodes to split. Cannot be smaller than 1.}

\item{minsplit}{(Only with survival outcome) Minimal number of events required to split the node. Cannot be smaller than 2.}

\item{cause}{(Only with competing events) Number indicates the event of interest.}

\item{IBS.min}{(Only with survival outcome) Minimal time to compute the Integrated Brier Score. Default value is set to 0.}

\item{IBS.max}{(Only with survival outcome) Maximal time to compute the Integrated Brier Score. Default value is set to the maximal time-to-event found.}

\item{seed}{Seed to replicate results}

\item{...}{Optional parameters to be passed to the low level function}
}
\value{
DynForest function return a list with the following elements:\tabular{ll}{
   \code{rf} \tab A table with each tree in column. Provide multiple charactistics about the tree building \cr
   \tab \cr
   \code{type} \tab Outcome type \cr
   \tab \cr
   \code{times} \tab A numeric vector containing the time-to-event for all subjects \cr
   \tab \cr
   \code{cause} \tab Indicating the cause of interest \cr
   \tab \cr
   \code{causes} \tab A numeric vector containing the causes indicator \cr
   \tab \cr
   \code{xerror} \tab A numeric vector containing the OOB error for each tree \cr
   \tab \cr
   \code{oob.err} \tab A numeric vector containing the OOB error for each subject \cr
   \tab \cr
   \code{oob.pred} \tab Outcome prediction for all subjects \cr
   \tab \cr
   \code{Importance} \tab A list of 3 elements: \code{Curve}, \code{Scalar} and \code{Factor}. Each element contains a numeric vector of VIMP statistic predictor in \code{Inputs} value \cr
   \tab \cr
   \code{gVIMP} \tab A numeric vector containing the gVIMP for each group defined in \code{imp.group} argument \cr
   \tab \cr
   \code{Inputs} \tab A list of 3 elements: \code{Curve}, \code{Scalar} and \code{Factor}. Each element contains the names of the predictors \cr
   \tab \cr
   \code{Curve.model} \tab A list of longitudinal markers containing the formula used for modeling in the random forest \cr
   \tab \cr
   \code{comput.time} \tab Computation time \cr
}
}
\description{
Built a random survival forest using multivariate longitudinal endogenous covariates (Devaux et al., 2022)
}
\details{
The function currently supports survival data (competing or single event).

FUTUR IMPROVEMENTS:
}
\examples{

}
\references{

}
\seealso{
\code{\link{predict_DynForest}}
}
\author{
Anthony Devaux (\email{anthony.devaux@u-bordeaux.fr})
}
