% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOB_tree.R
\name{OOB.tree}
\alias{OOB.tree}
\title{OOB tree}
\usage{
OOB.tree(
  tree,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  timeScale = 0.1,
  d_out = 0.1
)
}
\arguments{
\item{d_out}{}
}
\description{
OOB tree
}
\keyword{internal}
