% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_longitudinal_group.R
\name{permute_longitudinal_group}
\alias{permute_longitudinal_group}
\title{Permute a group of longitudinal markers between two subjects}
\usage{
permute_longitudinal_group(
  Longitudinal,
  id_var,
  time_var,
  marker_indices,
  idA,
  idB
)
}
\arguments{
\item{Longitudinal}{A list containing longitudinal data (as used in dynforest).}

\item{id_var}{Name of the subject identifier variable (e.g., "id").}

\item{time_var}{Name of the time variable (e.g., "time").}

\item{marker_indices}{Indices of the markers to permute.}

\item{idA}{The ID of the subject whose trajectories will be replaced.}

\item{idB}{The ID of the subject whose trajectories will be used for replacement.}
}
\value{
A data.frame containing the new trajectory for subject A.
}
\description{
This function swaps the trajectories of a set of longitudinal markers
from subject B into subject A while keeping the other markers for subject A unchanged.
}
