% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CIF.R
\name{plot_CIF}
\alias{plot_CIF}
\title{Plot the individual Cumulative Incidence Function (CIF) for the interest cause}
\usage{
plot_CIF(DynForestPred_obj, id = NULL)
}
\arguments{
\item{DynForestPred_obj}{An DynForestPred object from \code{predict()} function}

\item{id}{Identifiers for the selected subjects}
}
\value{
Display the CIF for selected subjects
}
\description{
Plot the individual Cumulative Incidence Function (CIF) for the interest cause
}
\examples{
\dontrun{
# Display CIF for subjects 102 and 260
plot_CIF(DynForestPred_obj = pred_dyn,
         id = c(102, 260))
}
}
