% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gVIMP.R
\name{plot_gVIMP}
\alias{plot_gVIMP}
\title{Compute the grouped-VIMP statistic}
\usage{
plot_gVIMP(DynForest_obj, PCT = FALSE)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}

\item{PCT}{Display VIMP statistic in pourcentage. Default value is FALSE.}
}
\value{

}
\description{
Compute the grouped-VIMP statistic
}
