% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gVIMP.R
\name{plot_gVIMP}
\alias{plot_gVIMP}
\title{Compute the grouped-VIMP statistic}
\usage{
plot_gVIMP(DynForest_obj, PCT = FALSE)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}

\item{PCT}{Display VIMP statistic in pourcentage. Default value is FALSE.}
}
\description{
Compute the grouped-VIMP statistic
}
\examples{
\dontrun{
data(pbc2)

# Build survival data
pbc2_surv <- unique(pbc2[,c("id","age","drug","sex","years","event")])

# Define time-independent continuous covariate
cont_covar <- list(X = pbc2_surv[,"age", drop = FALSE],
                   id = pbc2_surv$id)

# Define time-independent non continuous covariates
fact_covar <- list(X = pbc2_surv[,c("drug","sex")],
                   id = pbc2_surv$id)

# Define time-dependent continuous markers
cont_traj <- list(X = pbc2[,c("serBilir","SGOT","albumin","alkaline")],
                  id = pbc2$id,
                  time = pbc2$time,
                  model = list(serBilir = list(fixed = serBilir ~ time,
                                               random = ~ time),
                               SGOT = list(fixed = SGOT ~ time + I(time^2),
                                           random = ~ time + I(time^2)),
                               albumin = list(fixed = albumin ~ time,
                                              random = ~ time),
                               alkaline = list(fixed = alkaline ~ time,
                                               random = ~ time))
)

# Define outcome (survival here)
Y <- list(type = "surv",
          Y = Surv(pbc2_surv$years, factor(pbc2_surv$event)),
          id = pbc2_surv$id)

# Run DynForest function
res_dyn <- DynForest(Curve = cont_traj, Factor = fact_covar, Scalar = cont_covar,
                     Y = Y, ntree = 200, imp = TRUE,
                     imp.group = list(group1 = c("serBilir","SGOT"),
                                      group2 = c("albumin","alkaline")),
                     mtry = 3, nodesize = 2, minsplit = 3,
                     cause = 2, seed = 1234)

# Plot gVIMP
plot_gVIMP(res_dyn)

}
}
\seealso{
\code{\link{DynForest} \link{plot_VIMP} \link{plot_mindepth}}
}
