% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gVIMP.R
\name{plot_gVIMP}
\alias{plot_gVIMP}
\title{Compute the grouped-VIMP statistic}
\usage{
plot_gVIMP(DynForest_obj, PCT = FALSE)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}

\item{PCT}{Display VIMP statistic in pourcentage. Default value is FALSE.}
}
\description{
Compute the grouped-VIMP statistic
}
\examples{
\dontrun{

# Plot gVIMP
plot_gVIMP(res_dyn_gVIMP)

}
}
\seealso{
\code{\link{DynForest} \link{compute_gVIMP} \link{compute_VIMP} \link{plot_VIMP} \link{plot_mindepth}}
}
