% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mindepth.R
\name{plot_mindepth}
\alias{plot_mindepth}
\title{Compute the average minimal depth statistic}
\usage{
plot_mindepth(var_depth_obj, plot_level = c("predictor", "feature"))
}
\arguments{
\item{var_depth_obj}{Object from \code{var_depth} function}

\item{plot_level}{Compute the statistic at predictor (\code{plot_level}="predictor") or feature (\code{plot_level}="feature") level}
}
\description{
Compute the average minimal depth statistic
}
\examples{
\dontrun{

# Run var_depth function
res_varDepth <- var_depth(res_dyn)

}

}
\seealso{
\code{\link{DynForest} \link{var_depth} \link{plot_VIMP} \link{plot_gVIMP}}
}
