% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VIMP.R
\name{plot_VIMP}
\alias{plot_VIMP}
\title{Compute the VIMP statistic}
\usage{
plot_VIMP(DynForest_obj, PCT = FALSE, ordering = TRUE)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}

\item{PCT}{Display VIMP statistic in pourcentage. Default value is FALSE.}

\item{ordering}{Order predictors according to VIMP value. Default value is TRUE.}
}
\value{

}
\description{
Compute the VIMP statistic
}
