% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_DynForest.R
\name{predict.DynForest}
\alias{predict.DynForest}
\title{Predict with Dynamic random forests}
\usage{
\method{predict}{DynForest}(object, Curve = NULL, Scalar = NULL,
  Factor = NULL, timeScale = 0.1, ncores = NULL, ...)
}
\arguments{
\item{object}{: Dynamic random forest}

\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{timeScale}{[numeric]:}

\item{ncores}{[numeric]:}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
Predict with Dynamic random forests
}
