% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.DynForest}
\alias{predict.DynForest}
\title{Prediction using dynamic random forests}
\usage{
\method{predict}{DynForest}(
  DynForest_obj,
  timeData = NULL,
  fixedData = NULL,
  idVar,
  timeVar,
  t0 = NULL,
  ncores = NULL,
  parallel = TRUE,
  ...
)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} or \code{DynForest_OOB} object containing the dynamic random forest used on train data}

\item{t0}{Landmark time}

\item{ncores}{Number of cores used to grow trees in parallel. Default value is the number of cores of the computer-1.}

\item{parallel}{Allow paralleling. Default value is TRUE.}

\item{...}{optional parameters to be passed to the low level function}

\item{Curve}{A list of longitudinal predictors which should contain: \code{X} a dataframe with one row for repeated measurement and as many columns as markers; \code{id} is the vector of the identifiers for the repeated measurements contained in \code{X}; \code{time} is the vector of the measurement times contained in \code{X}.}

\item{Scalar}{A list of scalar predictors which should contain: \code{X} a dataframe with as many columns as scalar predictors; \code{id} is the vector of the identifiers for each individual.}

\item{Factor}{A list of factor predictors which should contain: \code{X} a dataframe with as many columns as factor predictors; \code{id} is the vector of the identifiers for each individual.}
}
\value{
Return the outcome of interest for the new subjects: matrix of probability of event of interest in survival mode, average value in regression mode and most likely value in classification mode
}
\description{
Prediction using dynamic random forests
}
\examples{
\dontrun{
data(pbc2)

# Build survival data
pbc2_surv <- unique(pbc2[,c("id","age","drug","sex","years","event")])

# Define time-independent continuous covariate
cont_covar <- list(X = pbc2_surv[,"age", drop = FALSE],
                   id = pbc2_surv$id)

# Define time-independent non continuous covariates
fact_covar <- list(X = pbc2_surv[,c("drug","sex")],
                   id = pbc2_surv$id)

# Define time-dependent continuous markers
cont_traj <- list(X = pbc2[,c("serBilir","SGOT","albumin","alkaline")],
                  id = pbc2$id,
                  time = pbc2$time,
                  model = list(serBilir = list(fixed = serBilir ~ time,
                                               random = ~ time),
                               SGOT = list(fixed = SGOT ~ time + I(time^2),
                                           random = ~ time + I(time^2)),
                               albumin = list(fixed = albumin ~ time,
                                              random = ~ time),
                               alkaline = list(fixed = alkaline ~ time,
                                               random = ~ time))
)

# Define outcome (survival here)
Y <- list(type = "surv",
          Y = Surv(pbc2_surv$years, factor(pbc2_surv$event)),
          id = pbc2_surv$id)

# Run DynForest function
res_dyn <- DynForest(Curve = cont_traj, Factor = fact_covar, Scalar = cont_covar,
                     Y = Y, ntree = 200, imp = TRUE,
                     imp.group = list(group1 = c("serBilir","SGOT"),
                                      group2 = c("albumin","alkaline")),
                     mtry = 3, nodesize = 2, minsplit = 3,
                     cause = 2, seed = 1234)

# Predict on new subjects using DynForest estimation (res_dyn object) from DynForest() function
pred_dyn <- predict(object = res_dyn,
                    Curve = cont_traj, Factor = fact_covar, Scalar = cont_covar,
                    t0 = 4)

}

}
