% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_shape_para.R
\name{rf_shape_para}
\alias{rf_shape_para}
\title{Parallelized Dynamic random Forest}
\usage{
rf_shape_para(Curve = NULL, Scalar = NULL, Factor = NULL,
  Shape = NULL, Image = NULL, Y, mtry, ntree, ncores,
  aligned.shape = FALSE, timeScale = 0.1, splitrule = NULL,
  nsplit_option = NULL, nodesize = 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Parallelized Dynamic random Forest
}
\keyword{internal}
