% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_shape_para.R
\name{rf_shape_para}
\alias{rf_shape_para}
\title{Parallelized Dynamic random Forest}
\usage{
rf_shape_para(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  mtry,
  ntree,
  ncores,
  timeScale = 0.1,
  nsplit_option = "quantile",
  nodesize = 1
)
}
\arguments{
\item{...}{}
}
\description{
Parallelized Dynamic random Forest
}
\keyword{internal}
