% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtmax.R
\name{Rtmax}
\alias{Rtmax}
\title{Randomized Dynamic tree}
\usage{
Rtmax(Curve = NULL, Scalar = NULL, Factor = NULL, Shape = NULL,
  Image = NULL, Y, mtry, aligned.shape = FALSE, timeScale = 0.1,
  splitrule = NULL, nsplit_option = NULL, nodesize = 1, ...)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{Y}{[list]:}

\item{mtry}{[integer]:}

\item{aligned.shape}{[logical]:}

\item{timeScale}{[numeric]:}

\item{...}{: option}
}
\description{
Randomized Dynamic tree
}
\keyword{internal}
