% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtmax_surv.R
\name{Rtmax_surv}
\alias{Rtmax_surv}
\title{Randomized Dynamic tree}
\usage{
Rtmax_surv(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y = NULL,
  mtry = 1,
  nsplit_option = "quantile",
  nodesize = 1,
  minsplit = 2,
  cause = 1,
  seed = 1234
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Y}{[list]:}

\item{mtry}{[integer]:}

\item{seed}{}
}
\description{
Randomized Dynamic tree
}
\keyword{internal}
