% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.DynForest}
\alias{summary.DynForest}
\title{Display the summary of DynForest}
\usage{
\method{summary}{DynForest}(object, ...)
}
\arguments{
\item{object}{\code{DynForest} or \code{DynForest_OOB} object}

\item{...}{Optional parameters to be passed to the low level function}
}
\value{
Return some information about the random forest
}
\description{
Display the summary of DynForest
}
\examples{
\dontrun{
# DynForest summary
summary(object = res_dyn_OOB)
}

}
\author{
Anthony Devaux (\email{anthony.devaux@u-bordeaux.fr})
}
