% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_depth.R
\name{var_depth}
\alias{var_depth}
\title{Extract characteristics from the trees building process}
\usage{
var_depth(DynForest_obj)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}
}
\value{

}
\description{
Extract characteristics from the trees building process
}
