% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_depth.R
\name{var_depth}
\alias{var_depth}
\title{Extract characteristics from the trees building process}
\usage{
var_depth(DynForest_obj)
}
\arguments{
\item{DynForest_obj}{\code{DynForest} object}
}
\value{
var_depth function return a list with the following elements:\tabular{ll}{
\code{min_depth} \tab A table providing for each feature in row: the average depth and the rank \cr
\tab \cr
\code{var_node_depth} \tab A table providing for each tree in column the minimal depth for each feature in row. NA indicates that the feature was not used for the corresponding tree \cr
\tab \cr
\code{var_count} \tab A table providing for each tree in column the number of times where the feature is used (in row). 0 value indicates that the feature was not used for the corresponding tree \cr
}
}
\description{
Extract characteristics from the trees building process
}
\examples{
\donttest{
data(pbc2)

# Sample 100 subjects
set.seed(1234)
id <- unique(pbc2$id)
id_sample <- sample(id, 100)
id_row <- which(pbc2$id\%in\%id_sample)

pbc2_train <- pbc2[id_row,]

timeData_train <- pbc2_train[,c("id","time",
                                "serBilir","SGOT",
                                "albumin","alkaline")]

# Create object with longitudinal association for each predictor
timeVarModel <- list(serBilir = list(fixed = serBilir ~ time,
                                     random = ~ time),
                     SGOT = list(fixed = SGOT ~ time + I(time^2),
                                 random = ~ time + I(time^2)),
                     albumin = list(fixed = albumin ~ time,
                                    random = ~ time),
                     alkaline = list(fixed = alkaline ~ time,
                                     random = ~ time))

# Build fixed data
fixedData_train <- unique(pbc2_train[,c("id","age","drug","sex")])

# Build outcome data
Y <- list(type = "surv",
          Y = unique(pbc2_train[,c("id","years","event")]))

# Run DynForest function
res_dyn <- DynForest(timeData = timeData_train, fixedData = fixedData_train,
                     timeVar = "time", idVar = "id",
                     timeVarModel = timeVarModel, Y = Y,
                     ntree = 50, nodesize = 5, minsplit = 5,
                     cause = 2, ncores = 1, seed = 1234)

# Run var_depth function
res_varDepth <- var_depth(res_dyn)

}

}
\seealso{
\code{\link{DynForest}}
}
