% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_split_MM.R
\name{var_split_MM}
\alias{var_split_MM}
\title{Mixed-Model Split function}
\usage{
var_split_MM(X, Y, timeScale = 0.1, nsplit_option = "quantile",
  nodesize = 1)
}
\arguments{
\item{nodesize}{}
}
\description{
Mixed-Model Split function
}
\keyword{internal}
