% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_split_MM.R
\name{var_split_MM}
\alias{var_split_MM}
\title{Mixed-Model Split function}
\usage{
var_split_MM(X, Y, nsplit_option = "quantile", cause = 1, init = NULL)
}
\arguments{
\item{init}{}
}
\description{
Mixed-Model Split function
}
\keyword{internal}
